Move inject methods into separate file.
